requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.55.0
exceptiongroup==1.1.3
fastapi==0.101.1
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
plantuml==0.3.0
pycparser==2.21
pydantic==2.2.0
pydantic_core==2.6.0
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
